shope/requirements.txt
python-dotenv==1.0.0
django-phonenumber-field==7.0.2
celery==5.2.7
redis==4.5.5
phonenumbers==8.13.11
django-compressor==4.4
jsmin==3.0.1
kombu==5.3.1
rjsmin==1.2.1
cssmin==0.2.0